send contexts up
