+++ packages/commons/search-utils/tsconfig.json
@@ -8,7 +8,7 @@
   "include": ["./src/**/*"],
   "references": [
     {
-      "path": "../../fdr-sdk"
+      "path": "../../fdr-sdk/tsconfig.build.json"
     },
     {
       "path": "../core-utils"
