+++ packages/fdr-sdk/tsconfig.json
@@ -7,7 +7,7 @@
     "lib": ["DOM", "ESNext"],
     "esModuleInterop": true
   },
-  "include": ["./src/**/*"],
+  "include": ["./src/**/*.{ts,tsx}"],
   "exclude": ["node_modules"],
   "references": [{ "path": "../commons/core-utils" }]
 }
