+++ packages/ui/app/src/atoms/types.ts
@@ -3,43 +3,11 @@ import type * as FernDocs from "@fern-api/fdr-sdk/docs";
 import type * as FernNavigation from "@fern-api/fdr-sdk/navigation";
 import { ColorsConfig, SidebarTab, VersionSwitcherInfo } from "@fern-ui/fdr-utils";
 import type { FernUser } from "@fern-ui/fern-docs-auth";
+import type { FeatureFlags } from "@fern-ui/fern-docs-utils";
 import { NextSeoProps } from "@fern-ui/next-seo";
 import { CustomerAnalytics } from "../analytics/types";
 import { DocsContent } from "../resolver/DocsContent";
 import { FernTheme } from "../themes/ThemedDocs";
-
-export interface FeatureFlags {
-    isApiPlaygroundEnabled: boolean;
-    isApiScrollingDisabled: boolean;
-    isWhitelabeled: boolean;
-    isSeoDisabled: boolean;
-    isTocDefaultEnabled: boolean;
-    isSnippetTemplatesEnabled: boolean;
-    isHttpSnippetsEnabled: boolean;
-    isInlineFeedbackEnabled: boolean;
-    isDarkCodeEnabled: boolean;
-    proxyShouldUseAppBuildwithfernCom: boolean;
-    isImageZoomDisabled: boolean;
-    useJavaScriptAsTypeScript: boolean;
-    alwaysEnableJavaScriptFetch: boolean;
-    scrollInContainerEnabled: boolean;
-    useMdxBundler: boolean;
-    isBatchStreamToggleDisabled: boolean;
-    isAuthEnabledInDocs: boolean;
-    isAiChatbotEnabledInPreview: boolean;
-    isAudioFileDownloadSpanSummary: boolean;
-    isDocsLogoTextEnabled: boolean;
-    isAudioExampleInternal: boolean;
-    usesApplicationJsonInFormDataValue: boolean;
-    isBinaryOctetStreamAudioPlayer: boolean;
-    hasVoiceIdPlaygroundForm: boolean;
-    isCohereTheme: boolean;
-    isFileForgeHackEnabled: boolean;
-    is404PageHidden: boolean;
-    isNewSearchExperienceEnabled: boolean;
-    // TODO: remove this after pinecone demo, this is a temporary flag
-    grpcEndpoints: readonly string[];
-}
 
 export interface NavigationProps {
     currentTabIndex: number | undefined;
