+++ packages/ui/app/src/index.ts
@@ -1,6 +1,6 @@
 export { type CustomerAnalytics } from "./analytics/types";
 export { DEFAULT_FEATURE_FLAGS } from "./atoms";
-export type { DocsProps, FeatureFlags } from "./atoms";
+export type { DocsProps } from "./atoms";
 export * from "./docs/DocsPage";
 export * from "./docs/NextApp";
 export { getApiRouteSupplier } from "./hooks/useApiRoute";
