+++ packages/ui/app/src/resolver/ApiDefinitionResolver.ts
@@ -1,9 +1,9 @@
 import type { DocsV1Read } from "@fern-api/fdr-sdk";
 import type * as FernNavigation from "@fern-api/fdr-sdk/navigation";
 import { isNonNullish, visitDiscriminatedUnion } from "@fern-ui/core-utils";
+import type { FeatureFlags } from "@fern-ui/fern-docs-utils";
 import { FernRegistry } from "../../../../fdr-sdk/src/client/generated";
 import { captureSentryErrorMessage } from "../analytics/sentry";
-import type { FeatureFlags } from "../atoms";
 import { type MDX_SERIALIZER } from "../mdx/bundler";
 import type { FernSerializeMdxOptions } from "../mdx/types";
 import { ApiEndpointResolver } from "./ApiEndpointResolver";
