+++ packages/ui/app/src/util/resolveDocsContent.ts
@@ -2,10 +2,10 @@ import type { APIV1Read, DocsV1Read } from "@fern-api/fdr-sdk/client/types";
 import type * as FernDocs from "@fern-api/fdr-sdk/docs";
 import * as FernNavigation from "@fern-api/fdr-sdk/navigation";
 import { isNonNullish, visitDiscriminatedUnion } from "@fern-ui/core-utils";
+import type { FeatureFlags } from "@fern-ui/fern-docs-utils";
 import GithubSlugger from "github-slugger";
 import { reverse } from "lodash-es";
 import { captureSentryError } from "../analytics/sentry";
-import type { FeatureFlags } from "../atoms";
 import { MDX_SERIALIZER } from "../mdx/bundler";
 import { getFrontmatter } from "../mdx/frontmatter";
 import type { FernSerializeMdxOptions } from "../mdx/types";
