+++ packages/ui/fern-docs-utils/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./constants";
+export * from "./flags";
 export * from "./getRedirectForPath";
 export * from "./isCustomDomain";
 export * from "./isDevelopment";
