+++ packages/ui/local-preview-bundle/tsconfig.json
@@ -14,7 +14,7 @@
       "path": "../../commons/fdr-utils"
     },
     {
-      "path": "../../fdr-sdk"
+      "path": "../../fdr-sdk/tsconfig.build.json"
     },
     {
       "path": "../app/tsconfig.build.json"
